avaliacao_final_marcos_paulo_nmap_default/requirements.txt
Flask==3.0.3
Jinja2==3.1.4
PyYAML==6.0.2
pydantic==2.9.2
rich==13.9.2
markdown-it-py==3.0.0
python-slugify==8.0.4
